Human patch from ctripcorp/apollo This commit is based on the source code from the following commit: http://github.com/ctripcorp/apollo/commit/a8a4a0d3f00e4193df90a76e20b4089574005d23 The mentioned commit triggered the following Travis build: http://travis-ci.org/ctripcorp/apollo/builds/444288772.

+++ apollo-adminservice/src/main/java/com/ctrip/framework/apollo/adminservice/controller/ClusterController.java
@@ -8,6 +8,7 @@ import com.ctrip.framework.apollo.common.exception.NotFoundException;
 import com.ctrip.framework.apollo.common.utils.BeanUtils;
 import com.ctrip.framework.apollo.common.utils.InputValidator;
 
+import com.ctrip.framework.apollo.core.ConfigConsts;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -51,10 +52,17 @@ public class ClusterController {
   @RequestMapping(path = "/apps/{appId}/clusters/{clusterName:.+}", method = RequestMethod.DELETE)
   public void delete(@PathVariable("appId") String appId,
                      @PathVariable("clusterName") String clusterName, @RequestParam String operator) {
+
     Cluster entity = clusterService.findOne(appId, clusterName);
+
     if (entity == null) {
       throw new NotFoundException("cluster not found for clusterName " + clusterName);
     }
+
+    if(ConfigConsts.CLUSTER_NAME_DEFAULT.equals(entity.getName())){
+      throw new BadRequestException("default cluster can not be delete.");
+    }
+
     clusterService.delete(entity.getId(), operator);
   }
